Move distribution of overlying grid data
