+++ edisgo/network/overlying_grid.py
@@ -3,10 +3,12 @@ from __future__ import annotations
 import logging
 import os
 
+from copy import deepcopy
 from zipfile import ZipFile
 
 import pandas as pd
 
+# from edisgo import EDisGo
 from edisgo.tools.tools import resample
 
 logger = logging.getLogger(__name__)
@@ -256,3 +258,165 @@ class OverlyingGrid:
 
         freq_orig = timeindex[1] - timeindex[0]
         resample(self, freq_orig, method, freq)
+
+
+def distribute_overlying_grid_requirements(edisgo_obj):
+    """
+    Distributes overlying grid requirements to components in grid.
+
+    Overlying grid requirements for e.g. electromobility charging are distributed to
+    all charging points where cars are parked, and for DSM to all DSM loads based
+    on their available load increase and decrease at each time step.
+
+    Parameters
+    -----------
+    edisgo_obj : :class:`~.EDisGo`
+        The eDisGo API object
+
+    Returns
+    --------
+    :class:`~.EDisGo`
+        New EDisGo object with only the topology data and adjusted time series data.
+
+    """
+
+    edisgo_copy = edisgo_obj.__class__()
+    edisgo_copy.topology = deepcopy(edisgo_obj.topology)
+    edisgo_copy.timeseries = deepcopy(edisgo_obj.timeseries)
+
+    # electromobility - distribute charging time series from overlying grid to all
+    # charging points based on upper power flexibility band
+    if not edisgo_obj.overlying_grid.electromobility_active_power.empty:
+        cp_loads = edisgo_obj.topology.loads_df.index[
+            edisgo_obj.topology.loads_df.type == "charging_point"
+        ]
+        # scale flexibility band upper power timeseries
+        scaling_df = edisgo_obj.electromobility.flexibility_bands[
+            "upper_power"
+        ].transpose() / edisgo_obj.electromobility.flexibility_bands["upper_power"].sum(
+            axis=1
+        )
+        edisgo_copy.timeseries._loads_active_power.loc[:, cp_loads] = (
+            scaling_df * edisgo_obj.overlying_grid.electromobility_active_power
+        ).transpose()
+
+    # storage units - distribute charging/discharging time series from overlying grid
+    # to all storage units based on their installed capacity
+    if not edisgo_obj.overlying_grid.storage_units_active_power.empty:
+        scaling_factor = (
+            edisgo_obj.topology.storage_units_df.p_nom
+            / edisgo_obj.topology.storage_units_df.p_nom.sum()
+        )
+        scaling_df = pd.DataFrame(
+            index=scaling_factor.index,
+            columns=edisgo_copy.timeseries.timeindex,
+            data=pd.concat(
+                [scaling_factor] * len(edisgo_copy.timeseries.timeindex), axis=1
+            ).values,
+        )
+        edisgo_copy.timeseries._storage_units_active_power = (
+            scaling_df * edisgo_obj.overlying_grid.storage_units_active_power
+        ).transpose()
+
+    # central PtH - distribute dispatch time series from overlying grid
+    # to all central PtH units based on their installed capacity
+    if not edisgo_obj.overlying_grid.heat_pump_central_active_power.empty:
+        hp_district = edisgo_obj.topology.loads_df[
+            (edisgo_obj.topology.loads_df.type == "heat_pump")
+            & (
+                edisgo_obj.topology.loads_df.sector.isin(
+                    ["district_heating", "district_heating_resistive_heater"]
+                )
+            )
+        ]
+        scaling_factor = hp_district.p_set / hp_district.p_set.sum()
+        scaling_df = pd.DataFrame(
+            index=scaling_factor.index,
+            columns=edisgo_copy.timeseries.timeindex,
+            data=pd.concat(
+                [scaling_factor] * len(edisgo_copy.timeseries.timeindex), axis=1
+            ).values,
+        )
+        edisgo_copy.timeseries._loads_active_power.loc[:, hp_district.index] = (
+            scaling_df
+            * edisgo_obj.overlying_grid.heat_pump_central_active_power.sum(axis=1)[0]
+        ).transpose()
+
+    # decentral PtH - distribute dispatch time series from overlying grid
+    # to all decentral PtH units based on their installed capacity
+    if not edisgo_obj.overlying_grid.heat_pump_decentral_active_power.empty:
+        hp_individual = edisgo_obj.topology.loads_df.index[
+            (edisgo_obj.topology.loads_df.type == "heat_pump")
+            & (
+                edisgo_obj.topology.loads_df.sector.isin(
+                    ["individual_heating", "individual_heating_resistive_heater"]
+                )
+            )
+        ]
+        # scale with heat pump upper power
+        scaling_factor = (
+            edisgo_obj.topology.loads_df.p_set.loc[hp_individual]
+            / edisgo_obj.topology.loads_df.p_set.loc[hp_individual].sum()
+        )
+        scaling_df = pd.DataFrame(
+            index=scaling_factor.index,
+            columns=edisgo_copy.timeseries.timeindex,
+            data=pd.concat(
+                [scaling_factor] * len(edisgo_copy.timeseries.timeindex), axis=1
+            ).values,
+        )
+        edisgo_copy.timeseries._loads_active_power.loc[:, hp_individual] = (
+            scaling_df * edisgo_obj.overlying_grid.heat_pump_decentral_active_power
+        ).transpose()
+
+    # DSM - distribute dispatch time series from overlying grid to all DSM loads based
+    # on their maximum load increase (in case of positive dispatch values) or
+    # their maximum load decrease (in case of negative dispatch values)
+    if not edisgo_obj.overlying_grid.dsm_active_power.empty:
+        dsm_loads = edisgo_obj.dsm.p_max.columns
+        if len(dsm_loads) > 0:
+            scaling_df_max = (
+                edisgo_obj.dsm.p_max.transpose() / edisgo_obj.dsm.p_max.sum(axis=1)
+            )
+            scaling_df_min = (
+                edisgo_obj.dsm.p_min.transpose() / edisgo_obj.dsm.p_min.sum(axis=1)
+            )
+            edisgo_copy.timeseries._loads_active_power.loc[:, dsm_loads] = (
+                edisgo_obj.timeseries._loads_active_power.loc[:, dsm_loads]
+                + (
+                    scaling_df_min
+                    * edisgo_obj.overlying_grid.dsm_active_power.clip(upper=0)
+                ).transpose()
+                + (
+                    scaling_df_max
+                    * edisgo_obj.overlying_grid.dsm_active_power.clip(lower=0)
+                ).transpose()
+            )
+        else:
+            logger.warning(
+                "EDisGo object has no attribute 'dsm'. DSM timeseries from "
+                "overlying grid cannot be distributed."
+            )
+
+    # curtailment
+    if not edisgo_obj.overlying_grid.renewables_curtailment.empty:
+        gens = edisgo_obj.topology.generators_df[
+            edisgo_obj.topology.generators_df.type.isin(["solar", "wind"])
+        ].index
+        gen_per_ts = edisgo_obj.timeseries.generators_active_power.loc[:, gens].sum(
+            axis=1
+        )
+        scaling_factor = (
+            edisgo_obj.timeseries.generators_active_power.loc[:, gens].transpose()
+            / gen_per_ts
+        ).fillna(0)
+        curtailment = (
+            scaling_factor * edisgo_obj.overlying_grid.renewables_curtailment
+        ).transpose()
+        edisgo_copy.timeseries._generators_active_power.loc[:, gens] = (
+            edisgo_obj.timeseries.generators_active_power.loc[:, gens] - curtailment
+        )
+
+    # reset reactive power time series
+    edisgo_copy.set_time_series_reactive_power_control()
+    return edisgo_copy
